商城发行、注册；平台注册；平台向司机转账

+++ src/main/java/org/fisco/bcos/business/entity/DriverEntity.java
@@ -26,7 +26,7 @@ public class DriverEntity {
     private long userId;
 
     @Column
-    private long platform;
+    private int platform;
 
     @Column
     private int star;
